HUB-981: Add support page with email
Added a new static support page to Self-Service instead of redirecting to an external page. The custom page is based on the page that we were redirecting to (https://www.verify.service.gov.uk/support/). The content has been cut down to only include information relevant to RPs or IDPs. A mailto link has been added so RPs and IDPs can find our support email address.

  - Added the `view_support` permission set to `true` - this allows all users to view the support page and only logged in users can see the page
  - Added a test for SupportController

  - Added the `-d` option to the startup script so the user can run the app with docker without providing input on the console
  - Updated README to provide instructions to log in when running locally

+++ startup.sh
@@ -1,21 +1,49 @@
 #!/bin/bash
 
-while true; do
-    read -p "Run using Docker? [y/n] " p
-    case $p in
-        y|Y|yes)    docker build . -f run.Dockerfile -t verify-self-service
-                    docker run --rm -p 8080:8080 -it verify-self-service
-                    break
-                    ;;
-        n|N|no)     set -e
-                    bundle check || bundle install
-                    yarn check || yarn install
-                    bin/rails s
-                    break
-                    ;;
-        x|X|exit)   exit 0
-                    ;;
-        *)          echo "Unknown option."
-                    ;;
-    esac
+RUN_IN_DOCKER=false
+
+show_help() {
+  cat << EOF
+Options:
+  -d , --docker   Run using docker
+EOF
+}
+
+while [ "$1" != "" ]; do
+  case $1 in
+      -d | --docker)    shift
+                        RUN_IN_DOCKER=true
+                        ;;
+      *)                echo -e "Unknown option $1...\n"
+                        show_help
+                        exit 1
+  esac
+  shift
 done
+
+if [[ $RUN_IN_DOCKER == 'false' ]]; then
+  while true; do
+      read -rp "Run using Docker? [y/n] " p
+      case $p in
+          y|Y|yes)    RUN_IN_DOCKER=true
+                      break
+                      ;;
+          n|N|no)     break
+                      ;;
+          x|X|exit)   exit 0
+                      ;;
+          *)          echo "Unknown option."
+                      ;;
+    esac
+  done
+fi
+
+if [[ $RUN_IN_DOCKER == 'true' ]]; then
+  docker build . -f run.Dockerfile -t verify-self-service
+  docker run --rm -p 8080:8080 -it verify-self-service
+else
+  set -e
+  bundle check || bundle install
+  yarn check || yarn install
+  bin/rails s
+fi
